Attempt to send some calls through Magnet

+++ magnet-npapi/src/npruntime.h
@@ -199,7 +199,7 @@ NP_END_MACRO
 #define STRINGZ_TO_NPVARIANT(_val, _v)                                        \
 NP_BEGIN_MACRO                                                                \
     (_v).type = NPVariantType_String;                                         \
-    NPString str = { _val, strlen(_val) };                                    \
+    NPString str = { _val, (int)strlen(_val) };                               \
     (_v).value.stringValue = str;                                             \
 NP_END_MACRO
